requirements.txt
# API and HTTP
requests==2.31.0
python-dotenv>=1.0.0

# LangChain and OpenAI
langchain-openai==0.0.5
httpx==0.27.0
tenacity==8.2.3
nest-asyncio==1.6.0

# Utilities
python-dateutil>=2.8.2

# Code Quality
ruff==0.4.4
black==24.4.2

# Testing & Evaluation
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
pytest==7.4.0
ragas==0.2.14
torch==2.1.0
transformers>=4.31.0
datasets>=2.14.6